Allowing to use multiple MultiMatchQuery when search by QS
Until now, we only had enabled a single `MultiMatchQuery` when try to search by a query string.

With this PR we've enabled to load multiple query string templates in order to do an `OR` between the multiple `MultiMatchQuery` queries. Also we now are loading properties as objects instead of load them as strings.

With this PR, we are now are going to improve our autocomplete using `synonyms` https://github.com/GrupoZapVivaReal/squad-search-ranking/issues/873